Ignore UserWarnings in some obvious pytest test cases.
The basic pytest often throw UserWarnings which refer to a meaningful scientific analysis rathter than code functionality. In these, known cases, the warnings get ignored. Also some minor ambiguities got fixed. Pytest completes without warnings now.

+++ pyplis/geometry.py
@@ -191,8 +191,8 @@ class MeasGeometry(object):
                    ("focal_length", float),  # in m
                    ("pix_width", float),  # in m
                    ("pix_height", float),  # in m
-                   ('pixnum_x', float),
-                   ('pixnum_y', float),
+                   ('pixnum_x', int),
+                   ('pixnum_y', int),
                    ('altitude_offs', float)])
 
     @property
